+++ CHANGELOG.md
@@ -22,3 +22,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Typography components: Title, Paragraph, Text
 - Add Accordion component
 - Add Input components: Input, Password, TextArea, File
+- Add Loader component